backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.1
mediapipe==0.10.21
opencv-python==4.8.1.78
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.2
flask-cors
requests
spacy
whisper
pdfplumber
